updates the citi python module
