+++ physionet-django/user/templates/user/citi/get_courses_by_member
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+    <soap12:Envelope xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+            xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:soap12="http://www.w3.org/2003/05/soap-envelope">
+      <soap12:Body>
+        <GetMemberCoursesbyID xmlns="https://webservices.citiprogram.org/">
+        <usr>{{username}}</usr>
+        <pwd>{{password}}</pwd>
+        <intMemberID>{{memberid}}</intMemberID>
+        </GetMemberCoursesbyID>
+      </soap12:Body>
+    </soap12:Envelope>
